refactor(transaction): remove broken methods, decouple cache, fix HandlingUseCase
- Remove doAfterCommitWithResult, doAfterRollbackWithResult, rollback()
  from TransactionOperationsOutputPort (broken or unused)
- Extract cache-invalidation-on-rollback into dedicated
  CacheInvalidationOnRollback component in cache adapter package
- Move CacheUtils to infrastructure.adapter.cache package
- Fix HandlingUseCase: tighten transactional boundary to persistence +
  event dispatch only; move value object validation outside transaction
  with direct presenter call (fixes silent error swallowing bug)
- Make doAfterRollback a no-op when no transaction is active
- Remove doAfterRollback mock stub from AbstractUseCaseTestSupport
