+++ src/main/java/com/github/cargoclean/core/port/transaction/TransactionOperationsOutputPort.java
@@ -63,37 +63,11 @@ public interface TransactionOperationsOutputPort {
     void doAfterCommit(TransactionRunnableWithoutResult runnableWithoutResult);
 
     /**
-     * Executes the provided runnable, returning the result of the execution, if called outside
-     * any transactional context. Otherwise, executes the runnable only after a successful
-     * commit of the current transaction.
-     *
-     * @param runnableWithResult a runnable with some return object
-     * @param <R>                any type
-     * @return object returned by the runnable
-     */
-    <R> R doAfterCommitWithResult(TransactionRunnableWithResult<R> runnableWithResult);
-
-    /**
-     * Executes the provided runnable if called outside any transactional context.
-     * Otherwise, executes the runnable only after the current transaction was rolled back.
+     * Executes the provided runnable only after the current transaction was rolled back.
+     * If called outside any transactional context, this method is a no-op: the runnable
+     * is not executed since there is no transaction to roll back.
      *
      * @param runnableWithoutResult a runnable which does not return anything
      */
     void doAfterRollback(TransactionRunnableWithoutResult runnableWithoutResult);
-
-    /**
-     * Executes the provided runnable, returning the result of the execution, if called outside
-     * any transactional context. Otherwise, executes the runnable only after the current
-     * transaction was rolled back.
-     *
-     * @param runnableWithResult a runnable with some return object
-     * @param <R>                any type
-     * @return object returned by the runnable
-     */
-    <R> R doAfterRollbackWithResult(TransactionRunnableWithResult<R> runnableWithResult);
-
-    /**
-     * Rolls back currently active transaction.
-     */
-    void rollback();
 }
